### 2019-06-21 #1515 @whedon
[PRE REVIEW]: modeLLtest: An R Package for Unbiased Model Comparison using Cross Validation
**Submitting author:** @ShanaScogin (<a href="http://orcid.org/0000-0002-7801-853X">Shana Scogin</a>)
**Repository:** <a href="https://github.com/ShanaScogin/modeLLtest" target ="_blank">https://github.com/ShanaScogin/modeLLtest</a>
**Version:** v1.0.0
