### 2019-06-21 #1515 @danielskatz
@ShanaScogin - you might want to change your subsection headings to ```##```s. If you do this, or make other changes, you can add a new comment in this issue containing ```@whedon generate pdf``` to see how it works out
