### 2019-06-24 2e0d93a
Fixed warnings from CMD check
