+++ tests/testthat/test_cvmf.R
@@ -1,4 +1,5 @@
 test_that("Simple model with two covariates runs with cvmf", {
+  skip_on_cran()
 
   set.seed(12345)
   x1 <- rnorm(100)
