+++ tests/testthat/test_cvll.R
@@ -1,4 +1,5 @@
 test_that("Simple model with cvll runs", {
+  skip_on_cran()
 
   set.seed(123456)
   b0 <- .2 # True value for the intercept
