+++ R/nepaldem_data.R
@@ -3,7 +3,9 @@
 #' Data from a study on the relationship bewteen land tenure and
 #' voter turnout in the three rounds of parliamentary elections
 #' in Nepal from the restoration of democracy in 1990 to 1999.
-#' Data is at the district-level (N = 75).
+#' Data is at the district-level (N = 75). Variable names are
+#' taken directly from original dataset. The data is publicly available
+#' and has been included here with the endorsement of the authors.
 #'
 #' @docType data
 #'
@@ -11,6 +13,7 @@
 #'
 #' @format A data frame with 76 rows and 73 variables:
 #' \describe{
+#'  \item{sn}{a column of identifiers. This column is not a variable}
 #'  \item{district}{names of the district in Nepal used in analysis}
 #'  \item{householdsize}{average size of household in district}
 #'  \item{total_holding}{total land holding}
