+++ tests/testthat/test_rlm_mm.R
@@ -1,4 +1,5 @@
 test_that("RLM-mm and cvdm with rlm-mm runs", {
+  skip_on_cran()
 
   data(nepaldem)
   nepaldem <- na.omit(nepaldem)
