+++ cran-comments.md
@@ -1,0 +1,10 @@
+## Test environments
+* local macOS High Sierra 10.13.6 install, R 3.5.1
+* ubuntu 16.04.6 LTS (on travis-ci), R 3.6.0
+* macOS High Sierra 10.13.3 (on travis-ci), R 3.6.0  
+
+## R CMD check results
+There were no ERRORs, WARNINGs or NOTEs
+
+## Downstream dependencies
+There are currently no downstream dependencies for this package.
