+++ man/cvll.Rd
@@ -33,7 +33,7 @@ used in the fit. All observations are included by default.}
 \item{na.action}{A missing-data filter function, applied to the model.frame,
 after any subset argument has been used.}
 
-\item{model}{The name of a statistical model to estimate.}
+\item{...}{Optional arguments, currently unsupported.}
 }
 \value{
 An object of class \code{cvll} computed by the cross-validated log likelihood
