### 2019-06-25 9d41e59
Updated description per cran
