### 2019-06-25 2c8247d
Updated cran comments
