+++ cran-comments.md
@@ -2,6 +2,7 @@
 * local macOS High Sierra 10.13.6 install, R 3.5.1
 * ubuntu 16.04.6 LTS (on travis-ci), R 3.6.0
 * macOS High Sierra 10.13.3 (on travis-ci), R 3.6.0  
+* win-builder (devel)
 
 ## R CMD check results
 There were no ERRORs, WARNINGs or NOTEs
