### 2019-06-25 1f6fb3e
Added more detail to description
