### 2019-06-26 b1fb99b
Updated description per cran suggestions
