### 2019-06-27 #1515 @usethedata
What say you @andrewheiss are you willing and able to review this paper for JOSS?  Thanks.
