### 2019-06-27 #1515 @usethedata
Also, @wrathematics @soodoku @lebebr01 @jacobsoj @mkearney -- would any of you be willing to review this paper for JOSS?  Thanks much.
