### 2019-06-30 a705155
Added rlm from mass to rcpp code
