+++ R/rlm_mm.R
@@ -1,200 +1,0 @@
-# from MASS Package
-# copyright (C) 1994-2016 W. N. Venables and B. D. Ripley
-# Comments from original code by Venebles and Ripley
-
-rlm_mm <-
-  function(x, y, weights, ..., w = rep(1, nrow(x)),
-           init = "ls", psi = psi.huber,
-           k2 = 1.345,
-           wt.method = c("inv.var", "case"),
-           maxit = 100, # changed from 20
-           acc = 1e-4,
-           test.vec = "resid",
-           lqs.control=NULL)
-  {
-    irls.delta <- function(old, new)
-      sqrt(sum((old - new)^2)/max(1e-20, sum(old^2)))
-    irls.rrxwr <- function(x, w, r)
-    {
-      w <- sqrt(w)
-      max(abs((matrix(r * w, 1L, length(r)) %*% x)/
-                sqrt(matrix(w, 1L, length(r)) %*% (x^2))))/sqrt(sum(w * r^2))
-    }
-    wmad <- function(x, w)
-    {
-      o <- sort.list(abs(x)); x <- abs(x)[o]; w <- w[o]
-      p <- cumsum(w)/sum(w)
-      n <- sum(p < 0.5)
-      if (p[n + 1L] > 0.5) x[n + 1L]/0.6745 else (x[n + 1L] + x[n + 2L])/(2*0.6745)
-    }
-
-    # method <- match.arg(method)
-    wt.method <- match.arg(wt.method)
-    nmx <- deparse(substitute(x))
-    if(is.null(dim(x))) {
-      x <- as.matrix(x)
-      colnames(x) <- nmx
-    } else x <- as.matrix(x)
-    if(is.null(colnames(x)))
-      colnames(x) <- paste("X", seq(ncol(x)), sep="")
-    if(qr(x)$rank < ncol(x))
-      stop("'x' is singular: singular fits are not implemented in 'rlm'")
-
-    if(!(any(test.vec == c("resid", "coef", "w", "NULL"))
-         || is.null(test.vec))) stop("invalid 'test.vec'")
-    ## deal with weights
-    xx <- x
-    yy <- y
-    if(!missing(weights)) {
-      if(length(weights) != nrow(x))
-        stop("length of 'weights' must equal number of observations")
-      if(any(weights < 0)) stop("negative 'weights' value")
-      if(wt.method == "inv.var") {
-        fac <- sqrt(weights)
-        y <- y*fac; x <- x* fac
-        wt <- NULL
-      } else {
-        w <- w * weights
-        wt <- weights
-      }
-    } else wt <- NULL
-
-      temp <- do.call("lqs",
-                      c(list(x, y, intercept = FALSE, method = "S",
-                             k0 = 1.548), lqs.control))
-      coef <- temp$coefficients
-      resid <- temp$residuals
-      psi <- psi.bisquare
-      if(length(arguments <- list(...)))
-        if(match("c", names(arguments), nomatch = 0L)) {
-          c0 <- arguments$c
-          if (c0 > 1.548) formals(psi)$c <- c0
-          else
-            warning("'c' must be at least 1.548 and has been ignored")
-        }
-      scale <- temp$scale
-
-    done <- FALSE
-    conv <- NULL
-    n1 <- (if(is.null(wt)) nrow(x) else sum(wt)) - ncol(x)
-    theta <- 2*pnorm(k2) - 1
-    gamma <- theta + k2^2 * (1 - theta) - 2 * k2 * dnorm(k2)
-
-    for(iiter in 1L:maxit) {
-      if(!is.null(test.vec)) testpv <- get(test.vec) ## check this
-
-      w <- psi(resid/scale)
-      if(!is.null(wt)) w <- w * weights
-      temp <- lm.wfit(x, y, w, method="qr")
[…]
