+++ NAMESPACE
@@ -5,6 +5,7 @@ export(cvll)
 export(cvlldiff)
 export(cvmf)
 importFrom(MASS,lqs)
+importFrom(MASS,rlm)
 importFrom(Rcpp,sourceCpp)
 importFrom(coxrobust,coxr)
 importFrom(quantreg,rq.fit)
