### 2019-07-02 5894e26
Updated examples per cran
