+++ R/cvll.R
@@ -30,7 +30,7 @@
 #'@return An object of class \code{cvll} computed by the cross-validated log likelihood
 #'(CVLL). See \code{cvdm_object} for more details.
 #'@examples
-#' \dontrun{
+#' \donttest{
 #'   set.seed(123456)
 #'   b0 <- .2 # True value for the intercept
 #'   b1 <- .5 # True value for the slope
