+++ R/cvlldiff.R
@@ -20,7 +20,7 @@
 #'Johnson's t-test. A positive test statistic supports the first method and a negative test
 #'statistic supports the second.See \code{cvdm_object} for more details.
 #'@examples
-#' \dontrun{
+#' \donttest{
 #'   set.seed(123456)
 #'   b0 <- .2 # True value for the intercept
 #'   b1 <- .5 # True value for the slope
