+++ man/cvll.Rd
@@ -50,7 +50,7 @@ Singular matrices during the leave-one-out cross-validation process
 are skipped.
 }
 \examples{
-\dontrun{
+\donttest{
   set.seed(123456)
   b0 <- .2 # True value for the intercept
   b1 <- .5 # True value for the slope
