+++ man/cvlldiff.Rd
@@ -36,7 +36,7 @@ statistic supports the method that produced the first vector and a negative
 test statistic supports the second.
 }
 \examples{
-\dontrun{
+\donttest{
   set.seed(123456)
   b0 <- .2 # True value for the intercept
   b1 <- .5 # True value for the slope
