### 2019-07-02 #1542 @whedon
[REVIEW]: modeLLtest: An R Package for Unbiased Model Comparison using Cross Validation
**Submitting author:** @ShanaScogin (<a href="http://orcid.org/0000-0002-7801-853X">Shana Scogin</a>)
**Repository:** <a href="https://github.com/ShanaScogin/modeLLtest" target ="_blank">https://github.com/ShanaScogin/modeLLtest</a>
**Version:** 1.0.1
