### 2019-07-02 #1515 @usethedata
@ShanaScogin Thanks for the assistance in finding reviewers.
@andrewheiss @monogan Thanks for agreeing to review.  There is a checklist for each of you over in the review issue ( #1542 ).  Please let me know if you have questions, either by tagging me in a post on that issue or reaching out to me directly.
