### 2019-07-02 d44c0e6
Updated journal title
