+++ JOSSpaper/ref.bib
@@ -8,7 +8,7 @@
 	language = {English (US)},
 	number = {4},
 	urldate = {2019-05-12},
-	journal = {Quality and Quantity},
+	journal = {Quality \& Quantity},
 	author = {Bruce A. Desmarais and Harden, Jeffrey J.},
 	month = jan,
 	year = {2014},
