### 2019-07-04 36c3ccd
Updated readme with CRAN how-to
