### 2019-08-04 #1542 @usethedata
Bumping issue.  I've reached out to @monogan and @andrewheiss via email.
