### 2019-08-05 #1542 @ShanaScogin
Thank you all for reviewing/editing, especially during the busy summer! Let me know if there are any questions or clarifications. We look forward to the review.
