### 2019-08-05 #1542 @andrewheiss
**modeLLtest** is a helpful new package that makes it easy to compare the performance of different models with cross-validated log-likelihood values. It includes two real-world datasets to accompany its examples, and it is fairly well documented. I came across a few issues, noted below. The biggest issue is that some objects are somewhat difficult to use since they're simple R lists without helper S3 functions, but this is easily addressed.

## Installation

The package installed without a problem on macOS, Linux, and Windows, both from source and from CRAN.

The README includes a note about gfortran-related installation errors. I wasn't able to get any of these to appear. It might be helpful to include the wording of the gfortran error in the README, or potentially include directions for fixing the error in the README in case the linked repository or the linked blog post go away in the future.

## Usage

The examples in the README and vignette run fine. It might be helpful to expand the examples a bit to (1) show how to extract the relevant information (even if it's just running `x$best` or `x$test_stat`, etc.), and (2) briefly show how to interpret the results and choose the best model. The JOSS paper helpfully goes over the technical details of why these tests and checks are important, but neither the paper nor the vignette/README show an example of how to choose which model is the best. Not knowing the statistical papers that the authors reference, I had to figure out which model was best by inspecting the resulting object and discovering the `$best`  slot. Without reviewing the referenced papers, I'm not quite sure how to interpret that—though I trust that it really is the best :)

Relatedly, it might be very helpful to provide helper functions for extracting the relevant information out of the `cvdm`-classed objects. This could be something like an S3 `print.cvdm` class that shows `$best`, `$test_stat`, and `$p_value` more clearly, or some other custom function. Right now, printing a `cvlldiff` object is okay, since it only has a few slots in the list and everything important is readable on one screen, but printing `cvdm` objects hides the relevant information since it's buried under less important list slots like `$call`, `$x`, and `$y`.

The duration of some models caught me by surprise. Running the earlier models in the README file (`obj_cvdm <- cvdm(Y ~ X, data.frame(cbind(Y, X)), method1 = "OLS", method2 = "MR")` for instance) was more or less instantaneous, but it took 100 seconds to run the `obj_cvmf_golder` `cvmf()` example. Noting that the model might take a while to fit in a comment in the example might be helpful. The JOSS paper states that "these tests now run in a reasonable amount of time on large data sets," but it'd be nice to know what "reasonable" means. Two minutes might be reasonable if these tests ordinarily take longer without the C++ code, but without a baseline, it's hard to tell.

## Documentation

It might be helpful to link to the documentation for `cvdm_object()` within the man pages for `cvdm()` and `cvll()`.

## Contributions

It might be helpful to have some community guidelines in the package too, such as a CONTRIBUTING.md file (perhaps modeled after something like [this](https://github.com/ropensci/nomisr/blob/master/CONTRIBUTING.md) or [this](https://github.com/tidyverse/ggplot2/blob/master/CONTRIBUTING.md)) and a CONDUCT.md file (like [this](https://github.com/ropensci/qualtRics/blob/master/CONDUCT.md) or [this](https://github.com/tidyverse/readr/blob/master/CONDUCT.md))

Creating an RStudio project `.Rproj` file could be helpful for encouraging package contributions, since it would enable RStudio's package building interface.

Right now, according to the README, bug tracking is handled via e-mail: "If you encounter a bug or have comments or questions, please email me at sscogin@nd.edu." It might be good to handle bugs more publicly through GitHub's issue tracker and include a `BugReports:` entry in `DESCRIPTION`. The startup message already directs users to the GitHub repository for development details—tracking issues there might be easiest and more community oriented.

## Paper

The JOSS paper is clear and easy to follow, establishes a clear need for the functions, and includes DOIs in all the references.

----

That's all I have. Great work!
