### 2019-08-06 081051e
Deleted unnec issue templates
