+++ .github/ISSUE_TEMPLATE/other-issue.md
@@ -1,10 +1,0 @@
----
-name: Other issue
-about: Describe this issue template's purpose here.
-title: ''
-labels: ''
-assignees: ''
-
----
-
-
