### 2019-08-06 b5a32b4
Added .Rproj and implemented issue handling per JOSS review
