### 2019-08-06 8f6ba70
updated documentation per JOSS review
