+++ R/cvlldiff.R
@@ -18,7 +18,7 @@
 #'@return An object of class \code{cvlldiff} computed by the cross-validated log likelihood
 #'difference in means test (CVDM). The test statistic object is the Cross-Validated
 #'Johnson's t-test. A positive test statistic supports the first method and a negative test
-#'statistic supports the second.See \code{cvdm_object} for more details.
+#'statistic supports the second.See \code{\link{cvdm_object}} for more details.
 #'@examples
 #' \donttest{
 #'   set.seed(123456)
