+++ man/cvmf.Rd
@@ -57,7 +57,7 @@ coefficients are treated as zeros.}
 \value{
 An object of class \code{cvmf} computed by the cross-validated median fit test
 (CVMF) to test between the PLM and IRR methods of estimating the Cox model.
-See \code{cvmf_object} for more details.
+See \code{\link{cvmf_object}} for more details.
 }
 \description{
 Applies cross-validated log-likelihood to test between
