+++ R/cvlldiff.R
@@ -12,6 +12,7 @@
 #'
 #'Desmarais, B. A., & Harden, J. J. (2014). An Unbiased Model Comparison Test Using
 #'Cross-Validation. Quality & Quantity, 48(4), 2155-2173.
+#'\href{https://doi.org/10.1007/s11135-013-9884-7}{https://doi.org/10.1007/s11135-013-9884-7}
 #'@param vector1 A numeric vector of cross-validated log-likelihoods.
 #'@param vector2 A numeric vector of cross-validated log-likelihoods.
 #'@param df A value of the degrees of freedom in the models.
@@ -19,6 +20,10 @@
 #'difference in means test (CVDM). The test statistic object is the Cross-Validated
 #'Johnson's t-test. A positive test statistic supports the first method and a negative test
 #'statistic supports the second.See \code{\link{cvdm_object}} for more details.
+#'@references
+#'Desmarais, B. A., & Harden, J. J. (2014). An Unbiased Model Comparison Test Using
+#'Cross-Validation. Quality & Quantity, 48(4), 2155-2173.
+#'\href{https://doi.org/10.1007/s11135-013-9884-7}{https://doi.org/10.1007/s11135-013-9884-7}
 #'@examples
 #' \donttest{
 #'   set.seed(123456)
