+++ man/cvll_object.Rd
@@ -19,12 +19,12 @@ The following components must be included in a legitimate \code{cvll} object.
 
 \item{method}{method of estimation.}
 
-The object also contain the following: \code{call}, \code{x}, and \code{y}. See \code{lm} documentation for more.
+The object also contain the following: \code{call}, \code{x}, and \code{y}. See \code{\link{lm}} documentation for more.
 
 }
 
 \seealso{
-\code{cvll}
+\code{\link{cvll}}
 }
 
 \keyword{cvll}
