### 2019-08-07 b5624d4
added code of conduct and contrib per joss
