+++ CONTRIBUTING.md
@@ -1,0 +1,14 @@
+# Contributing to `modeLLtest`
+
+We appreciate your interest in contributing to our project. This document contains guidelines for the contributing process. We welcome contributions from all, regardless of your level of experience.
+
+By participating in this project, you agree to abide by the [code of conduct](CODE_OF_CONDUCT.md).
+
+# Types of contributions 
+
+Contributions can come in many forms. This includes (but is not limited to!):
+
+- Identifying areas for future development ([open an Issue](https://github.com/ShanaScogin/modeLLtest/issues))
+- Identifying issues/bugs in the code ([open an Issue](https://github.com/ShanaScogin/modeLLtest/issues))
+- Adding functionality ([open a Pull Request](https://github.com/ShanaScogin/modeLLtest/pulls))
+- Fixing bugs ([open a Pull Request](https://github.com/ShanaScogin/modeLLtest/pulls))
