### 2019-08-07 #1542 @ShanaScogin
@andrewheiss Thank you for your time and comments! I've implemented the suggestions in the documentation and contributions sections and have begun addressing the suggestions in usage. Thank you again for your time - we appreciate the feedback!
