### 2019-08-10 #1542 @usethedata
@andrewheiss Thanks much for your thoughtful review.
