### 2019-08-20 #1542 @usethedata
@ShanaScogin Thanks for your patience.  In the interests of getting this completed, I went ahead and did the second review myself.  It was interesting to read through how you did the work.  It's been a while since I've done this sort of cross-validation work.  @andrewheiss did the hard work with the initial review and I don't have much to add.  I did, not surprisingly, have a bit of a challenge to install the current dev version, but no issues installing the CRAN version.  I don't think there's an action there for you, in that trying to anticipate all of the possible oddball problems that come up in compiling from source is not practical.  In my case, there was a path error that took a bit to find.

I've marked my review complete.
