### 2019-08-22 #1542 @ShanaScogin
Hi @usethedata thank you so much! We appreciate your review and will work on replicating those github download issues to make a more informative section in the readme as suggested. @monogan No worries - thank you for checking in, and hope the end of your summer is going well! I'm going to recompile the paper and deposit with Zenodo now. Thank you again to both reviewers for these wonderful comments - we look forward to incorporating them and getting an updated version up on CRAN!
