### 2019-08-22 #1542 @ShanaScogin
@usethedata I uploaded the pdf and the build to Zenodo - the doi is 10.5281/zenodo.3374054 and the version is 1.0.0. Is there something else I should do?
