### 2019-08-22 #1542 @usethedata
@ShanaScogin In your bib file, there's an item `@article{desmarais_testing_2013`. The DOI for that item is invalid.  I don't think you're actually referencing that article.  Can you either fix this DOI (if you're referencing the article and I'm not seeing it), or remove it from the .bib file?
