+++ JOSSpaper/ref.bib
@@ -15,22 +15,6 @@
 	pages = {2155--2173}
 }
 
-@article{desmarais_testing_2013,
-	title = {Testing for zero inflation in count models: {Bias} correction for the {Vuong} test},
-	volume = {13},
-	shorttitle = {Testing for zero inflation in count models},
-	url = {https://ageconsearch.umn.edu/record/252694},
-	doi = {st0319},
-	abstract = {The proportion of zeros in event-count processes may be inflated by an additional mechanism by which zeros are created. This has given rise to statistical models that accommodate zero inflation; these are available in Stata through the zip and zinb commands. The Vuong (1989, Econometrica 57: 307–333) test is regularly used to determine whether estimating a zero-inflation component is appropriate or whether a single-equation count model should be used. The use of the Vuong test in this case is complicated by the fact that zero-inflated models involve the estimation of several more parameters than the single-equation models. Although Vuong (1989, Econometrica 57: 307–333) suggested corrections to the test statistic to address the comparison of models with different numbers of parameters, Stata does not implement any such correction. The result is that the Vuong test used by Stata is biased toward supporting the model with a zero-inflation component, even when no zero inflation exists in the generative process. We provide new Stata commands for computing the Vuong statistic with corrections based on the Akaike and Bayesian (Schwarz) information criteria. In an extensive Monte Carlo study, we illustrate the bias inherent in using the uncorrected Vuong test, and we examine the relative merits of the Akaike and Schwarz corrections. Then, in an empirical example from international relations research, we show that errors in selecting an event-count model can have clear implications for substantive conclusions. Desmarais, Bruce A.; Harden, Jeffrey J.},
-	language = {en},
-	number = {4},
-	urldate = {2019-05-12},
-	journal = {Stata Journal},
-	author = {Desmarais, Bruce A. and Harden, Jeffrey J.},
-	year = {2013},
-	pages = {810}
-}
-
 @article{harden_linear_2011,
 	title = {Linear {Models} with {Outliers}: {Choosing} between {Conditional}-{Mean} and {Conditional}-{Median} {Methods}},
 	volume = {11},
