### 2019-08-22 #1542 @ShanaScogin
@usethedata - You're right. I removed it - do you want me to make an updated Zenodo with a new pdf, build, version number, and doi?
