### 2019-09-01 #1542 @labarba
@ShanaScogin — I don't see a tagged release in the software repository to match the version associated with the JOSS paper. Could you issue a release?
