### 2019-09-01 #1542 @labarba
Also, the final paragraph of your paper says that "the code is on GitHub" with a link. Please remove this. The link to the software repository and archive are both included in the margin decorators on the front page of the paper.
