### 2019-09-01 59132c8
removed sentence in paper per joss
