### 2019-09-01 #1542 @ShanaScogin
@labarba Thanks for the help. I've deleted that sentence and added release 1.0.1. Is there something else that needs to be done?
