### 2019-09-01 #1542 @labarba
Big thanks to our editor: @usethedata, and reviewers: @andrewheiss, @usethedata — thank you for your contributions to JOSS 🙏
